Refactor: 이미지 타입 파일 생성 및 이미지 사이즈 상수화
