+++ apps/web/app/(admin)/members/_components/sidepanel/ProfileImageUploader.tsx
@@ -3,22 +3,18 @@
 import { useState, useEffect, type ChangeEvent } from "react";
 import Image from "next/image";
 import DefaultProfileImage from "@ui/public/images/image_default_profile.png";
-import { type ImageFileType, type DisplayImageType, type FormImageType } from "@repo/types/src/membersType";
 import { MEMBER_FORM_MESSAGES } from "@repo/constants/messages";
-import { IMAGE_CONFIG } from "@repo/constants";
+import { IMAGE_CONFIG, IMAGE_SIZES } from "@repo/constants";
 import cn from "@ui/src/utils/cn";
+import { type ImageFileType, type DisplayImageType, type FormImageType, type ImageSize } from "@/app/types/ImageType";
 
 interface ProfileImageUploaderProps {
   currentImage: FormImageType;
   onImageChange?: (file: ImageFileType) => void;
-  size?: "sm" | "md";
+  size?: ImageSize;
 }
 
-export default function ProfileImageUploader({
-  currentImage,
-  onImageChange,
-  size = "md",
-}: ProfileImageUploaderProps): JSX.Element {
+export default function ProfileImageUploader({ currentImage, onImageChange }: ProfileImageUploaderProps): JSX.Element {
   const [imageObjectUrl, setImageObjectUrl] = useState("");
   const [isImageError, setIsImageError] = useState(false);
 
@@ -32,7 +28,7 @@ export default function ProfileImageUploader({
   };
 
   const getImageSource = (): DisplayImageType => {
-    if (isImageError) {
+    if (isImageError || !currentImage) {
       return DefaultProfileImage;
     }
 
@@ -40,11 +36,7 @@ export default function ProfileImageUploader({
       return imageObjectUrl;
     }
 
-    if (typeof currentImage === "string") {
-      return currentImage;
-    }
-
-    return DefaultProfileImage;
+    return currentImage;
   };
 
   const handleError = (): void => {
@@ -62,15 +54,12 @@ export default function ProfileImageUploader({
       <Image
         src={getImageSource()}
         alt={currentImage ? MEMBER_FORM_MESSAGES.IMAGE.PREVIEW_ALT : MEMBER_FORM_MESSAGES.IMAGE.DEFAULT_ALT}
-        width={size === "sm" ? 72 : 120}
-        height={size === "sm" ? 72 : 120}
+        width={IMAGE_SIZES.MD}
+        height={IMAGE_SIZES.MD}
         placeholder="blur"
         blurDataURL={IMAGE_CONFIG.BLUR_DATA_URL}
         onError={handleError}
-        className={cn("rounded-full object-cover", {
-          "size-72": size === "sm",
-          "size-120": size !== "sm",
-        })}
+        className={`rounded-full object-cover size-${IMAGE_SIZES.MD}`}
       />
       <label
         htmlFor="profileImage"
